s/uint16/uint32

+++ native-sql-engine/cpp/src/codegen/arrow_compute/ext/array_item_index.h
@@ -23,10 +23,10 @@ namespace codegen {
 namespace arrowcompute {
 namespace extra {
 struct ArrayItemIndex {
-  uint16_t id = 0;
-  uint16_t array_id = 0;
+  uint32_t id = 0;
+  uint32_t array_id = 0;
   ArrayItemIndex() : array_id(0), id(0) {}
-  ArrayItemIndex(uint16_t array_id, uint16_t id) : array_id(array_id), id(id) {}
+  ArrayItemIndex(uint32_t array_id, uint32_t id) : array_id(array_id), id(id) {}
 };
 }  // namespace extra
 }  // namespace arrowcompute
